Add check for if Lekko was introduced in current changes

+++ get_diff.sh
@@ -1,17 +1,28 @@
 #!/bin/bash -e
+
+# Expected args:
+# $1: base sha
+# $2: head sha
+# $3: whether head - base introduced Lekko to the repo
+
 git config --global user.email "108442683+lekko-app[bot]@users.noreply.github.com"
 git config --global user.name "lekko-app[bot]"
+# 000... sha means no previous commit
 if [[ "$1" != "0000000000000000000000000000000000000000" ]]; then
-  git reset --hard && git clean -fd
-  git checkout $1
-  echo "pre-sync base"
-  set +e
-  if ! lekko bisync -r ~/lekko; then
-    # If bisync fails on base commit, ignore and try to proceed
-    # This might result in detecting "unnecessary" additions, but hopefully fix PR should help resolve that
-    echo "Warning: bisync failed on base ($1). The current state of your Lekko repository will be used as the base of changes."
+  if [[ "$3" != true ]]; then
+    git reset --hard && git clean -fd
+    git checkout $1
+    echo "pre-sync base"
+    set +e
+    if ! lekko bisync -r ~/lekko; then
+      # If bisync fails on base commit, ignore and try to proceed
+      # This might result in detecting "unnecessary" additions, but hopefully fix PR should help resolve that
+      echo "Warning: bisync failed on base ($1). The current state of your Lekko repository will be used as the base of changes."
+    fi
+    set -e
+  else
+    echo "Base did not have Lekko"
   fi
-  set -e
 else
   echo "Base points to null commit"
 fi
